# doc/swag/swagger.yaml
basePath: /api/v1
definitions:
  github_com_sfshf_gonoweb_internal_app_web.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  github_com_sfshf_gonoweb_internal_model.TResource:
    properties:
      created_at:
        type: string
      deleted_at:
        type: integer
      icon:
        type: string
      id:
        type: integer
      identifier:
        type: string
      intro:
        type: string
      name:
        type: string
      type:
        type: integer
      updated_at:
        type: string
    type: object
...
  internal_app_web_ginsrv_resource.AddResourceReq:
    properties:
      icon:
        type: string
      id:
        type: string
      intro:
        type: string
      name:
        type: string
      type:
        enum:
        - 1
        - 2
        - 3
        type: integer
    type: object
  internal_app_web_ginsrv_resource.EditResourceReq:
    properties:
      icon:
        type: string
      intro:
        type: string
      name:
        type: string
    type: object
  internal_app_web_ginsrv_resource.ListResourceResp:
    properties:
      list:
        items:
          $ref: '#/definitions/github_com_sfshf_gonoweb_internal_model.TResource'
        type: array
      total:
        type: integer
    type: object
...
host: 192.168.211.128:8080
info:
  contact:
    email: Z1614789743@outlook.com
    name: sfshf
    url: https://github.com/sfshf
  description: This is a web api server.
  license:
    name: MIT
    url: https://github.com/sfshf/gonoweb/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Gono Web API
  version: "1.0"
paths:
...
  /resource:
    get:
      consumes:
      - text/plain
      description: 获取菜单/控件/API列表
      parameters:
      - description: 登录token
        in: header
        name: Authorization
        type: string
      - in: query
        name: id
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        maximum: 20
        minimum: 2
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_web_ginsrv_resource.ListResourceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
      summary: 获取菜单/控件/API列表
      tags:
      - 菜单/控件/API
    post:
      consumes:
      - application/json
      description: 新增菜单/控件/API
      parameters:
      - description: 登录token
        in: header
        name: Authorization
        type: string
      - description: 菜单/控件/API信息
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_app_web_ginsrv_resource.AddResourceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_model.TResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
      summary: 新增菜单/控件/API
      tags:
      - 菜单/控件/API
  /resource/:id:
    delete:
      consumes:
      - application/json
      description: 删除菜单/控件/API
      parameters:
      - description: 登录token
        in: header
        name: Authorization
        type: string
      - description: 菜单/控件/API的id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
      summary: 删除菜单/控件/API
      tags:
      - 菜单/控件/API
    get:
      consumes:
      - text/plain
      description: 获取菜单/控件/API信息
      parameters:
      - description: 登录token
        in: header
        name: Authorization
        type: string
      - description: 菜单/控件/API的id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_model.TResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
      summary: 获取菜单/控件/API信息
      tags:
      - 菜单/控件/API
    put:
      consumes:
      - application/json
      description: 编辑菜单/控件/API
      parameters:
      - description: 登录token
        in: header
        name: Authorization
        type: string
      - description: 菜单/控件/API的id
        in: path
        name: id
        type: string
      - description: 菜单/控件/API信息
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_app_web_ginsrv_resource.EditResourceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sfshf_gonoweb_internal_app_web.Response'
      summary: 编辑菜单/控件/API
      tags:
      - 菜单/控件/API
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"