# internal/model/t_resource.gen.go
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.

package model

import (
	"time"
)

const TableNameTResource = "t_resource"

// TResource mapped from table <t_resource>
type TResource struct {
	ID         int64     `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
	CreatedAt  time.Time `gorm:"column:created_at;default:CURRENT_TIMESTAMP" json:"created_at"`
	UpdatedAt  time.Time `gorm:"column:updated_at;default:CURRENT_TIMESTAMP" json:"updated_at"`
	DeletedAt  int64     `gorm:"column:deleted_at" json:"deleted_at"`
	Type       int32     `gorm:"column:type;not null" json:"type"`
	Identifier string    `gorm:"column:identifier;not null" json:"identifier"`
	Name       string    `gorm:"column:name;not null" json:"name"`
	Intro      string    `gorm:"column:intro" json:"intro"`
	Icon       string    `gorm:"column:icon" json:"icon"`
}

// TableName TResource's table name
func (*TResource) TableName() string {
	return TableNameTResource
}